requirements.txt
apache-airflow
apache-airflow-providers-http
apache-airflow-providers-postgres
pandas
numpy
wget
fastapi
SQLAlchemy